+++ src/example_serialization/feature_test/save_load.cpp
@@ -30,8 +30,8 @@ namespace for_feature_test
 	{
 		std::printf("Result:\n");
 		std::printf("  Fatal-error:    %s\n", has_fatal_error ? "Occured!" : "nothing");
-		std::printf("  Save data size: %d\n", save_data_size);
-		std::printf("  Used work-buffer size(peak): %d\n", peak_work_size);
+		std::printf("  Save data size: %d\n", static_cast<int>(save_data_size));
+		std::printf("  Used work-buffer size(peak): %d\n", static_cast<int>(peak_work_size));
 	}
 	void printResultForOutput(const archive::results& result)
 	{
